# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HookCreateRequest": {
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "environment": {
            "additionalProperties": true,
            "title": "Environment",
            "type": "object"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "script_path": {
            "title": "Script Path",
            "type": "string"
          },
          "timeout": {
            "default": 30,
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Timeout",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "event",
          "script_path"
        ],
        "title": "HookCreateRequest",
        "type": "object"
      },
      "HookResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "hook_id": {
            "title": "Hook Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "script_path": {
            "title": "Script Path",
            "type": "string"
          },
          "timeout": {
            "title": "Timeout",
            "type": "integer"
          }
        },
        "required": [
          "hook_id",
          "name",
          "event",
          "script_path",
          "enabled",
          "timeout",
          "created_at"
        ],
        "title": "HookResponse",
        "type": "object"
      },
      "ScheduleCreateRequest": {
        "properties": {
          "category": {
            "default": "all",
            "title": "Category",
            "type": "string"
          },
          "interval_minutes": {
            "default": 60,
            "maximum": 10080.0,
            "minimum": 5.0,
            "title": "Interval Minutes",
            "type": "integer"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "name",
          "query"
        ],
        "title": "ScheduleCreateRequest",
        "type": "object"
      },
      "ScheduleUpdateRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "interval_minutes": {
            "anyOf": [
              {
                "maximum": 10080.0,
                "minimum": 5.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interval Minutes"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ScheduleUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Search and merge results across multiple torrent trackers",
    "title": "\u0411\u043e\u0431\u0430 Search Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hooks": {
      "get": {
        "operationId": "list_hooks_api_v1_hooks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Hooks",
        "tags": [
          "hooks"
        ]
      },
      "post": {
        "operationId": "create_hook_api_v1_hooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Hook",
        "tags": [
          "hooks"
        ]
      }
    },
...
    "/api/v1/hooks/{hook_id}": {
      "delete": {
        "operationId": "delete_hook_api_v1_hooks__hook_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "hook_id",
            "required": true,
            "schema": {
              "title": "Hook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Hook",
        "tags": [
          "hooks"
        ]
      }
    },
...
    "/api/v1/schedules": {
      "get": {
        "operationId": "list_schedules_api_v1_schedules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Schedules",
        "tags": [
          "scheduler"
        ]
      },
      "post": {
        "operationId": "create_schedule_api_v1_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Schedule",
        "tags": [
          "scheduler"
        ]
      }
    },
    "/api/v1/schedules/{schedule_id}": {
      "delete": {
        "operationId": "delete_schedule_api_v1_schedules__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Schedule",
        "tags": [
          "scheduler"
        ]
      },
      "get": {
        "operationId": "get_schedule_api_v1_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedule",
        "tags": [
          "scheduler"
        ]
      },
      "patch": {
        "operationId": "update_schedule_api_v1_schedules__schedule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Schedule",
        "tags": [
          "scheduler"
        ]
      }
    },
...
  }
}